Make the fix for bug 448943 work.
2008-04-04  Matthias Clasen  <mclasen@redhat.com>

        * glib/gmain.c: Make the fix for bug 448943 work.



svn path=/branches/glib-2-16/; revision=6827

+++ glib/gmain.c
@@ -3320,7 +3320,7 @@ g_timeout_set_expiration (GTimeoutSource *timeout_source,
       if (!session_bus_address)
         session_bus_address = g_getenv ("HOSTNAME");
       if (session_bus_address)
-        timer_perturb = ABS (g_str_hash (session_bus_address));
+        timer_perturb = ABS ((gint) g_str_hash (session_bus_address));
       else
         timer_perturb = 0;
     }
